Fixed Structure for Testing
- Fixed folder structure to run tests
- Added two sample tests that will always pass
- Edited travis file to automatically execute tests
